removed unnecessary cast

+++ ev3classes/src/lejos/hardware/sensor/MindSensorsPressureSensor.java
@@ -101,7 +101,7 @@ private class PressureMode implements SensorMode {
 	@Override
 	public void fetchSample(float[] sample, int offset) {
 		getData(0x53, buf, 0, 4);		
-		sample[offset] = (float) EndianTools.decodeIntLE(buf, 0);
+		sample[offset] = EndianTools.decodeIntLE(buf, 0);
 	}
 
 	@Override
